Tesseract.PixBox — Type

struct PixBox
    x::Int32
    y::Int32
    w::Int32
    h::Int32
end

The PixBox structure is used to identify a region in an image.

Values:

Name	Description

x	The left side of the region identified by this box.

y	The top edge of the region identified by this box.

w	The width of the box.

h	The height of the box.