Tesseract.pix_read_tiff — Function

pix_read_tiff(
    filename::AbstractString;
    page::Integer = Int32(1)
)::Union{Pix, Nothing}

Read a TIFF image from the specified file. Returns nothing on error.

Arguments:

T	Name	Default	Description

R	filename		The name of the TIFF file to load.

O	page	Int32(1)	The image to load from the file.

Restrictions:

page - Must be greater than 0.

Details:

TIFF files can contain multiple images. The page parameter allows you to specify which image you want to load.

pix_read_tiff(
    stream::IO;
    page::Integer = Int32(1)
)::Union{Pix, Nothing}

Read a TIFF image from the specified stream. Returns nothing on error.

Arguments:

T	Name	Default	Description

R	stream		The IO stream to read the TIFF file from.

O	page	Int32(1)	The image to load from the file.

Restrictions:

page - Must be greater than 0.

Details:

This implementation mirrors the API provided by Leptonica when you pass in a FILE pointer. This assumes that the remainder of the stream contains a TIFF image.

TIFF files can contain multiple images. The page parameter allows you to specify which image you want to load.

pix_read_tiff(
    data::AbstractArray{UInt8};
    page::Integer = Int32(1)
)::Union{Pix, Nothing}

Read a TIFF image from the byte array. Returns nothing on error.

Arguments:

T	Name	Default	Description

R	data		The byte array to read the TIFF image from.

O	page	Int32(1)	The image to load from the file.

Restrictions:

page - Must be greater than 0.

Details:

TIFF files can contain multiple images. The page parameter allows you to specify which image you want to load.