Tesseract.pix_write_pnm — Function

pix_write_pnm(
    filename::AbstractString,
    pix::Pix
)::Bool

Write an image to a file in the PNM image format. If there is an error false is returned.

Arguments:

T	Name	Default	Description

R	filename		The name of the file to write to.

R	pix		The image to write to the file.

Details:

If the file exists it will be overwritten.

pix_write_pnm(
    stream::IO,
    pix::Pix
)::Bool

Write an image to an IO stream in the PNM image format. If there is an error false is returned.

Arguments:

T	Name	Default	Description

R	stream		The IO stream to write the image to.

R	pix		The image to write to the IO stream.

pix_write_pnm(
    pix::Pix
)::Union{Vector{UInt8}, Nothing}

Write an image to a byte array in the PNM image format. If there is an error false is returned.

Arguments:

T	Name	Default	Description

R	pix		The image to write to a byte array.