Tesseract.tess_get_param — Function

tess_get_param(
    inst::TessInst,
    name::AbstractString,
    ::Type{T}
)::Union{T, Nothing} where T<:Integer

Retrieve an integer parameter from the Tesseract engine. Returns nothing if the value could not be retrieved.

Arguments:

T	Name	Default	Description

R	inst		The instance to read the parameter from.

R	name		The name of the parameter to read.

R	::Type{T}		The type to return.

Examples:

julia> using Tesseract

julia> download_languages()
true

julia> instance = TessInst()
Allocated Tesseract instance.

julia> tess_get_param(instance, "edges_min_nonhole", Int)
12

See also: tess_params, tess_params_parsed, tess_set_param

tessgetparam( inst::TessInst, name::AbstractString, ::Type{Bool} )::Union{Bool, Nothing}

Retrieve a boolean parameter from the Tesseract engine. Returns nothing if the value could not be retrieved.

Arguments:

T	Name	Default	Description

R	inst		The instance to read the variable from.

R	name		The name of the variable to read.

R	::Type{Bool}		Identifies that you want a boolean value.

Examples:

julia> using Tesseract

julia> download_languages()
true

julia> instance = TessInst()
Allocated Tesseract instance.

julia> tess_get_param(instance, "edges_debug", Bool)
false

See also: tess_params, tess_params_parsed, tess_set_param

tess_get_param(
    inst::TessInst,
    name::AbstractString,
    ::Type{Float64}
)::Union{Float64, Nothing}

Retrieve a Float parameter from the Tesseract engine. Returns nothing if the value could not be retrieved.

Arguments:

T	Name	Default	Description

R	inst		The instance to read the variable from.

R	name		The name of the parameter to read.

R	::Type{Float64}		Identifies that you want a Float64 value.

Examples:

julia> using Tesseract

julia> download_languages()
true

julia> instance = TessInst()
Allocated Tesseract instance.

julia> tess_get_param(instance, "classify_min_slope", Float64)
0.414213562

See also: tess_params, tess_params_parsed, tess_set_param

tessgetparam( inst::TessInst, name::AbstractString, ::Type{String} )::Union{String, Nothing}

Retrieve a String parameter from the Tesseract engine. Returns nothing if the value could not be retrieved.

Arguments:

T	Name	Default	Description

R	inst		The instance to read the variable from.

R	name		The name of the parameter to read.

R	::Type{Float64}		Identifies that you want a Float64 value.

Examples:

julia> using Tesseract

julia> download_languages()
true

julia> instance = TessInst()
Allocated Tesseract instance.

julia> tess_get_param(instance, "page_separator", String)
"\f"

See also: tess_params, tess_params_parsed, tess_set_param