Tesseract.TessInst — Type

mutable struct TessInst
    ptr::Ptr{Cvoid}
end

A wrapper for the Api object in the Tesseract library.

Values:

Name	Description

ptr	The pointer to the Api object allocated by the C library.

Details:

Most method calls cannot use this object until tess_init() called on it for initialization.

When the garbage collector collects this object the associated pointer object will be freed in the library. The object can also be manually freed by calling tess_delete!() on it.

See also: TessInst(languages::AbstractString, dataPath::AbstractString).

Tesseract.TessInst — Method

TessInst(
    languages::AbstractString = "eng",
    dataPath::AbstractString = TESS_DATA
)

Construct an initialize a TessInst object.

Arguments:

T	Name	Default	Description

O	languages	eng	The language(s) to load.

O	dataPath	tessdata	The directory to look for the language files in.

Details:

To change the langauges identified by this instance you can call tess_init(). Multiple langagues can be specified by seperating them with a plus. So if you want english and spanish you could specify "eng+spa". The language codes are (usually) the ISO 639-3 code.

Example:

julia> using Tesseract

julia> download_languages("eng+fra")
true

julia> instance = TessInst("eng+fra")
Allocated Tesseract instance.

See also: tess_init.