Tesseract.tess_params — Function

tess_params(
    inst::TessInst,
    filename::AbstractString
)::Bool

Print out all the parameters with their values and help text to the specified file. Returns false if there was an error.

Arguments:

T	Name	Default	Description

R	inst		The Tesseract instance to get get the parameters from.

R	filename		The filename to write to.

Details:

For each parameter this method prints out it's name, it's value, and some descriptive text about the variable. Each variable is on it's own line with a tab character seperating each value.

See also: tess_params_parsed, tess_get_param, tess_set_param

tess_params(
    inst::TessInst,
    stream::IO
)::Bool

Print out all the parameters with their values and help text to the specified stream. Returns false if there was an error.

Arguments:

T	Name	Default	Description

R	inst		The Tesseract instance to get get the parameters from.

R	stream		The stream to write the files to.

Details:

For each parameter this method prints out it's name, it's value, and some descriptive text about the variable. Each variable is on it's own line with a tab character seperating each value.

See also: tess_params_parsed, tess_get_param, tess_set_param

tess_params(
    inst::TessInst
)::Union{String, Nothing}

Print out all the parameters with their values and help text to a string. Returns nothing if there was an error.

Arguments:

T	Name	Default	Description

R	inst		The Tesseract instance to get the parameters from.

Details:

The return string will contain multiple lines, each line contains the name of a variable, it's value, and some descriptive text about the variable. The fields are separated by tabs.

See also: tess_params_parsed, tess_get_param, tess_set_param